+++ internal/runner/progress.go
@@ -3,10 +3,14 @@ package runner
 import (
 	"github.com/mih-kopylov/bulker/internal/config"
 	"github.com/schollz/progressbar/v3"
+	"github.com/sirupsen/logrus"
 )
 
 type Progress interface {
+	// Incr increase current progress with 1
 	Incr()
+	// IndicateTermination makes progress indicate that bulker received "SIGINT" and terminates gracefully
+	IndicateTermination()
 }
 
 func NewProgress(conf *config.Config, maxCount int) Progress {
@@ -18,7 +22,14 @@ type ProgressBarProgress struct {
 }
 
 func (p *ProgressBarProgress) Incr() {
-	_ = p.bar.Add(1)
+	err := p.bar.Add(1)
+	if err != nil {
+		logrus.Debug(err)
+	}
+}
+
+func (p *ProgressBarProgress) IndicateTermination() {
+	p.bar.Describe("terminating, please wait")
 }
 
 func newProgressBarProgress(maxCount int, visible bool) *ProgressBarProgress {